+++ MonoGame.Framework/Graphics/PackedVector/HalfVector4.cs
@@ -51,7 +51,7 @@ namespace Microsoft.Xna.Framework.Graphics.PackedVector
         /// <returns>The ulong containing the packed values.</returns>
         static ulong PackHelper(ref Vector4 vector)
         {
-            ulong num4 = (ulong)HalfTypeHelper.Convert(vector.X);
+            ulong num4 = HalfTypeHelper.Convert(vector.X);
             ulong num3 = ((ulong)HalfTypeHelper.Convert(vector.Y) << 0x10);
             ulong num2 = ((ulong)HalfTypeHelper.Convert(vector.Z) << 0x20);
             ulong num1 = ((ulong)HalfTypeHelper.Convert(vector.W) << 0x30);
